+++ src/app/record/components/affiliation-stacks-groups/modals/modal-affiliations/modal-affiliations.component.ts
@@ -158,21 +158,38 @@ export class ModalAffiliationsComponent implements OnInit, OnDestroy {
 
     this.affiliationForm = this._formBuilder.group(
       {
-        organization: new UntypedFormControl(this.organization, {
-          validators: [
-            Validators.required,
-            Validators.maxLength(MAX_LENGTH_LESS_THAN_ONE_THOUSAND),
-          ],
-        }),
-        city: new UntypedFormControl(this.city, {
-          validators: [
-            Validators.required,
-            Validators.maxLength(MAX_LENGTH_LESS_THAN_ONE_THOUSAND),
-          ],
-        }),
-        region: new UntypedFormControl(this.region, {
-          validators: [Validators.maxLength(MAX_LENGTH_LESS_THAN_ONE_THOUSAND)],
-        }),
+        organization: new UntypedFormControl(
+          {
+            value: this.organization,
+            disabled: this.selectedOrganizationFromDatabase,
+          },
+          {
+            validators: [
+              Validators.required,
+              Validators.maxLength(MAX_LENGTH_LESS_THAN_ONE_THOUSAND),
+            ],
+          }
+        ),
+        city: new UntypedFormControl(
+          { value: this.city, disabled: this.selectedOrganizationFromDatabase },
+          {
+            validators: [
+              Validators.required,
+              Validators.maxLength(MAX_LENGTH_LESS_THAN_ONE_THOUSAND),
+            ],
+          }
+        ),
+        region: new UntypedFormControl(
+          {
+            value: this.region,
+            disabled: this.selectedOrganizationFromDatabase,
+          },
+          {
+            validators: [
+              Validators.maxLength(MAX_LENGTH_LESS_THAN_ONE_THOUSAND),
+            ],
+          }
+        ),
         country: new UntypedFormControl('', {
           validators: [Validators.required],
         }),
@@ -279,6 +296,9 @@ export class ModalAffiliationsComponent implements OnInit, OnDestroy {
               (x) => x.value === this.affiliation.country.value
             ).key,
           })
+          if (this.selectedOrganizationFromDatabase) {
+            this.affiliationForm.get('country').disable()
+          }
         }
       })