+++ CHANGELOG.md
@@ -1,3 +1,23 @@
+## v2.88.0 - 2024-04-28
+
+[Full Changelog](https://github.com/ORCID/orcid-angular/compare/v2.87.0...v2.88.0)
+
+### Fix
+
+- Read only organization field when editing affiliation if selected from db
+
+## v2.87.0 - 2024-04-28
+
+[Full Changelog](https://github.com/ORCID/orcid-angular/compare/v2.86.0...v2.87.0)
+
+### Fix
+
+- Add missing hint to country and city when a disambiguated org been has been selected (#2243)
+
+## v2.86.0 - 2024-04-28
+
+[Full Changelog](https://github.com/ORCID/orcid-angular/compare/v2.85.0...v2.86.0)
+
 ## v2.85.0 - 2024-04-23
 
 [Full Changelog](https://github.com/ORCID/orcid-angular/compare/v2.84.0...v2.85.0)